refactor: move splitter weights to view; change developerview to viewbase
